refactor: create first class collection
