+++ src/main/java/kitchenpos/domain/TableGroup.java
@@ -1,17 +1,11 @@
 package kitchenpos.domain;
 
-import org.springframework.util.CollectionUtils;
-
-import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 @Entity
 public class TableGroup {
@@ -20,8 +14,7 @@ public class TableGroup {
     private Long id;
     private LocalDateTime createdDate;
 
-    @OneToMany(mappedBy = "tableGroup", cascade = CascadeType.PERSIST)
-    private List<OrderTable> orderTables = new ArrayList<>();
+    private OrderTables orderTables;
 
     protected TableGroup() {
     }
@@ -31,33 +24,16 @@ public class TableGroup {
     }
 
     public TableGroup(Long id, LocalDateTime createdDate, List<OrderTable> orderTables) {
-        validateOrderTables(orderTables);
         this.id = id;
         this.createdDate = createdDate;
-        this.orderTables = new ArrayList<>(orderTables);
-    }
-
-    private void validateOrderTables(List<OrderTable> orderTables) {
-        if (CollectionUtils.isEmpty(orderTables) || orderTables.size() < 2) {
-            throw new IllegalArgumentException("주문 테이블은 2개 이상이여야 합니다");
-        }
+        this.orderTables = new OrderTables(orderTables);
     }
 
     public void changeOrderTables(List<OrderTable> orderTables) {
-        validateEmptyTables(orderTables);
-        validateOrderTables(orderTables);
         for (OrderTable orderTable : orderTables) {
             orderTable.changeTableGroup(this);
         }
-        this.orderTables = new ArrayList<>(orderTables);
-    }
-
-    private void validateEmptyTables(List<OrderTable> orderTables) {
-        for (final OrderTable savedOrderTable : orderTables) {
-            if (!savedOrderTable.isEmpty() || Objects.nonNull(savedOrderTable.getTableGroup())) {
-                throw new IllegalArgumentException("단체 지정은 빈 테이블만 가능합니다");
-            }
-        }
+        this.orderTables = new OrderTables(orderTables);
     }
 
     public Long getId() {
@@ -69,6 +45,6 @@ public class TableGroup {
     }
 
     public List<OrderTable> getOrderTables() {
-        return orderTables;
+        return orderTables.getOrderTables();
     }
 }
